pyproject.toml
[tool.poetry]
name = "playwright_exp"
version = "0.1.0"
description = ""
authors = ["fuffes <fuffes05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.33.0"
pytest-playwright = "^0.3.3"
pytest-asyncio = "^0.21.0"
asyncio = "^3.4.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
